blono_dev_server/celery.py
from __future__ import absolute_import, unicode_literals
import os
import dotenv

from celery import Celery
from django.conf import settings

dotenv.read_dotenv(
    os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env')
)

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'blono_dev_server.settings')

app = Celery('blono_dev_server')
app.config_from_object('django.conf:settings', namespace='CELERY')
app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)

@app.task(bind=True)
def debug_task(self):
    print('Request: {0!r}'.format(self.request))
