requirements.txt
amqp==2.4.2
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.3
Babel==2.6.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.2
django-admin==1.3.2
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==2.5.2
django-excel-response2==2.0.8
django-six==1.0.4
django-timezone-field==3.0
djangorestframework==3.9.2
flower==0.9.3
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
kombu==4.5.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.1
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.3.1
pylint-django==2.0.7
pylint-plugin-utils==0.5
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
screen==1.0.1
six==1.12.0
slackclient==1.3.1
sqlparse==0.3.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.4
urllib3==1.24.2
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.54.0
wrapt==1.11.1
xlwt==1.3.0
